:basketball:

+++ leetcode/algorithms/IsomorphicStrings.java
@@ -27,8 +27,8 @@ public class IsomorphicStrings {
         int[] array1 = new int[128];
         int[] array2 = new int[128];
         for (int i = 0; i < s.length(); i++) {
-            int c1 = (int) s.charAt(i);
-            int c2 = (int) t.charAt(i);
+            int c1 = s.charAt(i);
+            int c2 = t.charAt(i);
             if (array1[c1] != 0 && array1[c1] != c2 ||
                     array2[c2] != 0 && array2[c2] != c1)
                 return false;
